% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitME.R
\name{fitME}
\alias{fitME}
\title{An ME model-fitting function}
\usage{
fitME(X, Y, u, q, weights = rep(1, nrow(X)), error.type = "clasical")
}
\arguments{
\item{X}{design matrix of independent variable values (variables contained in columns)}

\item{Y}{vector of dependent variable values}

\item{u}{design matrix of measurement error standard deviations on X (first column, corresponding to the intercept term, is usually a vector of 0s). This should have the same dimensions as X with corresponding elements}

\item{q}{vector of measurement error standard deviations on Y}

\item{weights}{optional weights for each observation in (X,Y)}

\item{error.type}{classical or non-classical measurement error model}
}
\description{
This function lets you fit a regression model incorporating measurement error on independent or dependent variables
}
\examples{
fitME()
}
\keyword{error,}
\keyword{measurement}
\keyword{regression}
\keyword{sampling}
