% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.kernel.R
\name{gaussian.kernel}
\alias{gaussian.kernel}
\title{A function to calculate Gaussian kernel-based weights for geographically weighted regression}
\usage{
gaussian.kernel(distance, bandwidth)
}
\arguments{
\item{distance}{matrix of distances between locations in a defined spatial area}

\item{bandwidth}{bandwidth value used in calculating Gaussian weights; bandwidth selected using standard AIC or CV error minimization}
}
\description{
This function lets you calculate distance-based observation weghts using a Gaussian kernel
}
\examples{
gaussian.kernel()
}
\keyword{geographically}
\keyword{regression}
\keyword{weighted}
