% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialME.R
\name{spatialME}
\alias{spatialME}
\title{A GWR model-fitting function adjusting for measurement error}
\usage{
spatialME(X, Y, u, q, DM, bw, p.ci = 0.95, error.type = "classical")
}
\arguments{
\item{X}{design matrix of independent variable values (variables contained in columns)}

\item{Y}{vector of dependent variable values}

\item{u}{design matrix of measurement error standard deviations on X (first column, corresponding to the intercept term, is usually a vector of 0s). This should have the same dimensions as X with corresponding elements}

\item{q}{vector of measurement error standard deviations on Y}

\item{DM}{n x n distance matrix (matrix of distances between each n locations)}

\item{p.ci}{percent confidence intervals}

\item{error.type}{classical or non-classical measurement error model}
}
\description{
This function lets you fit a geographically weighted regression model incorporating measurement error on independent or dependent variables that may vary spatially
}
\examples{
spatialME()
}
\keyword{error,}
\keyword{geographically}
\keyword{measurement}
\keyword{regression}
\keyword{sampling}
\keyword{weighted}
